pyproject.toml
[project]
name = "methane-super-emitters"
version = "0.0.1"
authors = [
  { name="Vytautas Jancauskas", email="vytautsa.jancauskas@dlr.de" },
]
description = "Methane super emitter detection using AI"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pytest",
    "click",
    "pdoc",
    "matplotlib",
    "geopandas",
    "geopy",
    "scipy",
    "netCDF4",
    "numpy",
    "shapely",
    "rasterio",
    "geedim",
    "joblib",
    "lightning",
    "torchvision",
    "cdsapi>=0.7.2",
    "black",
]

[project.urls]
Homepage = "https://github.com/DLR-MF-DAS/methane-super-emitters"
Issues = "https://github.com/DLR-MF-DAS/methane-super-emitters/issues"
